pyproject.toml
[project]
name = "friendsapp"
version = "0.1.0"
description = "Tinder-like app with Profile Management Service"
authors = [
    {name = "Zoey Zuo",email = "zoeyzuouk@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=4.2.0,<5.0.0",
    "djangorestframework>=3.14.0,<4.0.0",
    "psycopg2-binary>=2.9.5,<3.0.0",
    "python-dotenv>=0.21.0,<1.0.0",
    "django-storages>=1.13.2,<2.0.0",
    "azure-storage-blob>=12.14.1,<13.0.0",
    "dj-database-url>=1.2.0,<2.0.0",
    "gunicorn>=20.1.0,<21.0.0",
    "whitenoise>=6.3.0,<7.0.0",
    "pillow (>=11.2.1,<12.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
